+++ src/ll/misc.py
@@ -83,66 +83,61 @@ except ImportError:
 			return string
 
 
-# Use Python implementations of those functions
-_defaultitem = object()
-
-def item(iterator, index, default=_defaultitem):
+def item(iterable, index, default=None):
 	"""
 	Returns the :obj:`index`'th element from the iterable. :obj:`index` may be
 	negative to count from the end. E.g. 0 returns the first element produced
 	by the iterator, 1 the second, -1 the last one etc. If :obj:`index` is
 	negative the iterator will be completely exhausted, if it's positive it
 	will be exhausted up to the :obj:`index`'th element. If the iterator
-	doesn't produce that many elements :exc:`IndexError` will be raised,
-	except when :obj:`default` is given, in which case :obj:`default` will be
-	returned.
+	doesn't produce that many elements :obj:`default` will be returned.
 	"""
 	i = index
 	if i>=0:
-		for item in iterator:
+		for item in iterable:
 			if not i:
 				return item
 			i -= 1
 	else:
 		i = -index
 		cache = collections.deque()
-		for item in iterator:
+		for item in iterable:
 			cache.append(item)
 			if len(cache)>i:
 				cache.popleft()
 		if len(cache)==i:
 			return cache.popleft()
-	if default is _defaultitem:
-		raise IndexError(index)
-	else:
-		return default
+	return default
 
 
-def first(iterator, default=_defaultitem):
+def first(iterable, default=None):
 	"""
 	Return the first element from the iterable. If the iterator doesn't
-	produce any elements :exc:`IndexError` will be raised, except when
-	:obj:`default` is given, in which case :obj:`default` will be returned.
+	produce any elements :obj:`default` will be returned.
 	"""
-	return item(iterator, 0, default)
+	for item in iterable:
+		return item
+	return default
 
 
-def last(iterator, default=_defaultitem):
+def last(iterable, default=None):
 	"""
 	Return the last element from the iterable. If the iterator doesn't produce
-	any elements :exc:`IndexError` will be raised, except when :obj:`default`
-	is given, in which case :obj:`default` will be returned.
+	any elements :obj:`default` will be returned.
 	"""
-	return item(iterator, -1, default)
+	item = default
+	for item in iterable:
+		pass
+	return item
 
 
-def count(iterator):
+def count(iterable):
 	"""
 	Count the number of elements produced by the iterable. Calling this
 	function will exhaust the iterator.
 	"""
 	count = 0
-	for node in iterator:
+	for node in iterable:
 		count += 1
 	return count