+++ src/ll/oradd.py
@@ -42,6 +42,7 @@ The content of the generated file ``data.oradd`` will look like this::
 	{'name': 'person_insert', 'keys': ['per_id'], 'args': {'per_id': 'max', 'per_firstname': 'Max', 'per_lastname': 'Mustermann'}}
 	{'name': 'contact_insert', 'keys': ['per_id', 'con_id'], 'args': {'per_id': 'max', 'con_id': 'max_mail', 'con_type': 'email', 'con_value': 'max@example.org'}}
 	{'type': 'file', 'name': 'protrait_{max}.png', 'content': b'\x89PNG\r\n\x1a\n...'}
+	{'type': 'resetsequence', 'sequence': 'person_seq', 'table': 'person', 'field': 'per_id'}
 
 i.e. it's just one Python ``repr`` of a dictionary per line.
 
@@ -70,6 +71,8 @@ calling ``contact_insert``. The PL/SQL equivalent of the above is::
 		)
 	end;
 
+Furthermore it will copy one file and reset the sequence ``person_seq``.
+
 
 Data format
 -----------
@@ -93,7 +96,8 @@ printed for display purposes, the original format is on one line)::
 The keys in this dictionary have the following meaning:
 
 	``type`` : string (optional)
-		This is either ``"procedure"`` (the default) or ``"file"``.
+		This is either ``"procedure"`` (the default), ``"file"`` or
+		``"resetsequence"``.
 
 	``name`` : string (required)
 		The name of the procedure to be called or the name of the file to be
@@ -117,7 +121,7 @@ For type ``"procedure"`` the following additional keys are used:
 		key (i.e. a key that has the same identifier) oradd will pass the value
 		from the first use as a normal ``IN`` parameter. The ``keys`` key is
 		optional, without it no parameter will be treated as a key. ``keys`` may
-		also be a dictionary, wich the identifiers as the keys and values the names
+		also be a dictionary, with the identifiers as the keys and values the names
 		of Python types. This can be used for out parameters that are not integers.
 
 	``sqls`` : list (optional)
@@ -131,6 +135,27 @@ For type ``"file"`` the following additional key is used:
 
 	``content``: bytes (required)
 		The content of the file to be created.
+
+For type ``"resetsequence"`` the following additional key is used:
+
+	``sequence``: string (required)
+		The name of the sequence to reset. The sequence will be reset to the
+		maximum value of a field in a table.
+
+	``table``: string (required)
+		The name of the table that contains the field.
+
+	``field``: string (required)
+		The name of the field in the table ``table``. The sequence will be reset
+		to a value, so that fetching the next value from the sequence will deliver
+		a value that is larger than the maximum value of the field ``field`` in the
+		table ``table``.
+
+	``minvalue``: integer (optional, default 10)
+		The minimum value for the sequence.
+
+	``increment``: integer (optional, default 10)
+		The increment (i.e. the steop size) for the sequence.
 
 An oradd file in UL4ON format contains the same dictionaries, but not as a
 Python repr output, but in UL4ON format. The UL4ON dump is *not* a list of
@@ -163,6 +188,11 @@ it supports the following command line options:
 		call), ``2`` (like ``1``, plus a summary of which procedure has been
 		called how often), ``3`` (detailed output for each procedure call, plus
 		summary)
+
+	``-c``, ``--commit``
+		Specifies when to commit database transactions. ``record`` commit after
+		every procedure call. ``once`` at the end of the script and ``never`` rolls
+		back the transaction after all imports.
 """
 
 # We're importing ``datetime``, so that it's available to ``eval()``
@@ -320,9 +350,9 @@ def _formatcall(record, allkeys):
 			if argvalue is None:
 				args.append("{}=None".format(argname))
 			elif argvalue in allkeys:
-				args.append("{}={}".format(argname, allkeys[argvalue]))
+				args.append("{}={}=<{}>".format(argname, allkeys[argvalue], argvalue))
 			else:
-				args.append("{}=?".format(argname))
+				args.append("{}=<{}>".format(argname, argvalue))
 		elif argname in sqls:
 			args.append("{}={}".format(argname, argvalue))
 		else:
@@ -391,13 +421,25 @@ def copyfile(name, content, allkeys):
 		os.remove(tempname)
 
 
+def resetsequence(cursor, sequence, table, field, minvalue, increment):
+	cursor.execute("select nvl(max({}), {}) from {}".format(field, minvalue, table))
+	tabvalue = cursor.fetchone()[0]
+	cursor.execute("select {}.nextval from dual".format(sequence))
+	seqvalue = cursor.fetchone()[0]
+	cursor.execute("alter sequence {} increment by {}".format(sequence, max(minvalue, tabvalue-seqvalue)))
+	cursor.execute("select {}.nextval from dual".format(sequence))
+	seqvalue = cursor.fetchone()[0]
+	cursor.execute("alter sequence {} increment by {}".format(sequence, increment))
+	return seqvalue
+
+
 def main(args=None):
 	p = argparse.ArgumentParser(description="Import an oradd dump to an Oracle database", epilog="For more info see http://www.livinglogic.de/Python/oradd/index.html")
 	p.add_argument("connectstring", help="Oracle connect string")
 	p.add_argument("file", nargs="?", help="Name of dump file (default: read from stdin)", type=argparse.FileType("r"), default=sys.stdin)
 	p.add_argument("-f", "--format", dest="format", help="Format of the dumpfile ('oradd' or 'ul4on') (default %(default)s)", default="oradd", choices=("oradd", "ul4on"))
 	p.add_argument("-v", "--verbose", dest="verbose", help="Give a progress report? (default %(default)s)", type=int, default=2, choices=(0, 1, 2, 3))
-	p.add_argument("-r", "--rollback", dest="rollback", help="Do a rollback after all imports instead of a commit? (default %(default)s)", default=False, action="store_true")
+	p.add_argument("-c", "--commit", dest="commit", help="When should database transactions be committed? (default %(default)s)", default="once", choices=("record", "once", "never"))
 
 	args = p.parse_args(args)
 
@@ -407,6 +449,7 @@ def main(args=None):
 		allkeys = {}
 		counts = collections.Counter()
 		countfiles = 0
+		countsequences = 0
 		loader = dict(oradd=load_oradd, ul4on=load_ul4on)[args.format]
 		cursor = db.cursor()
 		for (i, record) in enumerate(loader(args.file), 1):
@@ -415,13 +458,19 @@ def main(args=None):
 				if args.verbose >= 3:
 					if type == "procedure":
 						sys.stdout.write("#{}: procedure {}".format(i, _formatcall(record, allkeys)))
-					else:
+					elif type == "file":
 						sys.stdout.write("#{}: file {}".format(i, record["name"].format(**allkeys)))
+					elif type == "resetsequence":
+						sys.stdout.write("#{}: resetting sequence {} to maximum value from {}.{}".format(i, record["sequence"], record["table"], record["field"]))
+					else:
+						raise ValueError("unknown command type {!r}".format(type))
 				else:
 					sys.stdout.write(".")
 				sys.stdout.flush()
 			if type == "procedure":
 				newkeys = importrecord(record, cursor, allkeys)
+				if args.commit == "record":
+					db.commit()
 				if args.verbose >= 3:
 					if newkeys:
 						sys.stdout.write(" -> {}\n".format(", ".join("{}={!r}".format(argname, argvalue) for (argname, argvalue) in newkeys.items())))
@@ -429,24 +478,30 @@ def main(args=None):
 						sys.stdout.write("\n")
 				sys.stdout.flush()
 				counts[record["name"]] += 1
-			else:
+			elif type == "file":
 				copyfile(record["name"], record["content"], allkeys)
 				if args.verbose >= 3:
 					sys.stdout.write(" -> {} bytes written\n".format(len(record["content"])))
 					sys.stdout.flush()
 				countfiles += 1
-		if args.rollback:
-			db.rollback()
-		else:
+			elif type == "resetsequence":
+				newvalue = resetsequence(cursor, sequence=record["sequence"], table=record["table"], field=record["field"], minvalue=record.get("minvalue", 10), increment=record.get("increment", 10))
+				if args.verbose >= 3:
+					sys.stdout.write(" -> reset to {}\n".format(newvalue))
+					sys.stdout.flush()
+				countsequences += 1
+		if args.commit == "once":
 			db.commit()
+		elif args.commit == "never":
+			db.rollback()
 	finally:
 		if args.verbose >= 3:
 			print()
 
 	if args.verbose >= 2:
 		totalcount = sum(counts.values())
-		l1 = len(str(max(totalcount, countfiles)))
-		l2 = max(len(procname) for procname in counts)
+		l1 = len(str(max(totalcount, countfiles, countsequences)))
+		l2 = max(len(procname) for procname in counts) if counts else 0
 		print()
 		print("Summary")
 		print("="*(l1+1+l2))
@@ -457,6 +512,7 @@ def main(args=None):
 		print("{} {}".format("-"*l1, "-"*l2))
 		print("{:>{}} (total calls)".format(totalcount, l1))
 		print("{:>{}} (files)".format(countfiles, l1))
+		print("{:>{}} (sequences)".format(countsequences, l1))
 
 
 if __name__ == "__main__":